remove extra folder from temporal types tests
