+++ packages/graphql/tests/api-v6/integration/create/types/localDateTime.int.test.ts
@@ -18,10 +18,10 @@
  */
 
 import neo4jDriver from "neo4j-driver";
-import type { UniqueType } from "../../../../../utils/graphql-types";
-import { TestHelper } from "../../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../utils/graphql-types";
+import { TestHelper } from "../../../../utils/tests-helper";
 
-describe("Create Nodes with LocalTime fields", () => {
+describe("Create Nodes with LocalDateTime fields", () => {
     const testHelper = new TestHelper({ v6Api: true });
     let Movie: UniqueType;
 
@@ -29,7 +29,7 @@ describe("Create Nodes with LocalTime fields", () => {
         Movie = testHelper.createUniqueType("Movie");
         const typeDefs = /* GraphQL */ `
         type ${Movie.name} @node {
-            localTime: LocalTime
+            localDateTime: LocalDateTime
         }
     `;
         await testHelper.initNeo4jGraphQL({ typeDefs });
@@ -38,22 +38,22 @@ describe("Create Nodes with LocalTime fields", () => {
     afterEach(async () => {
         await testHelper.close();
     });
-
-    test("should be able to create nodes with LocalTime fields", async () => {
+    
+    test("should be able to create nodes with LocalDateTime fields", async () => {
         const time1 = new Date("2024-02-17T11:49:48.322Z");
         const time2 = new Date("2025-02-17T12:49:48.322Z");
 
-        const neoTime1 = neo4jDriver.LocalTime.fromStandardDate(time1);
-        const neoTime2 = neo4jDriver.LocalTime.fromStandardDate(time2);
+        const neoTime1 = neo4jDriver.LocalDateTime.fromStandardDate(time1);
+        const neoTime2 = neo4jDriver.LocalDateTime.fromStandardDate(time2);
 
         const mutation = /* GraphQL */ `
             mutation {
                 ${Movie.operations.create}(input: [
-                        { node: { localTime: "${neoTime1.toString()}" } }
-                        { node: { localTime: "${neoTime2.toString()}" } }
+                        { node: { localDateTime: "${neoTime1.toString()}" } }
+                        { node: { localDateTime: "${neoTime2.toString()}" } }
                     ]) {
                     ${Movie.plural} {
-                        localTime
+                        localDateTime
                     }
                 }
             }
@@ -65,8 +65,8 @@ describe("Create Nodes with LocalTime fields", () => {
         expect(gqlResult.data).toEqual({
             [Movie.operations.create]: {
                 [Movie.plural]: expect.toIncludeSameMembers([
-                    { localTime: neoTime1.toString() },
-                    { localTime: neoTime2.toString() },
+                    { localDateTime: neoTime1.toString() },
+                    { localDateTime: neoTime2.toString() },
                 ]),
             },
         });
